Moving to more library-like structure
