+++ numcmctools/__init__.py
@@ -1,0 +1,2 @@
+from numcmctools.plotstack import PlotStack
+from numcmctools.plot import Plot
