+++ numcmctools/plot.py
